\name{PPI}
\alias{PPI}
\docType{data}
\title{The protein protein interaction network data}
\description{A two column matrix containing with each
row containing the names of the interacting gene pairs}
\usage{data(PPI)}
\format{
  The format is:
 int [1:1000, 1:2] 184 270 85 386 11 302 4 42 173 233 ...
}
\examples{
data(PPI)
}
\keyword{datasets}
