\name{scaled_edge_score}
\alias{scaled_edge_score}
\docType{data}
\title{The scaled ECF statistics of all the edges}
\description{The scaled ECF statistics measuring 
the differential correlation of all the edges}
\usage{data(scaled_edge_score)}
\format{
  The format is:
 num [1:1000] 3.1315 -0.6512 -0.8184 -0.596 -0.0935 ...
}
\examples{
data(scaled_edge_score)
}
\keyword{datasets}
