\name{scaled_node_score}
\alias{scaled_node_score}
\docType{data}
\title{The scaled ECF-statistics of all the edges}
\description{The scaled ECF-statistics measuring the
differential correlation of all the edges}
\usage{data(scaled_node_score)}
\format{
  The format is:
 Named num [1:500] -0.416 -0.392 -0.326 -0.335 0.237 ...
 - attr(*, "names")= chr [1:500] "1" "2" "3" "4" ...
}
\examples{
data(scaled_node_score)
}
\keyword{datasets}
