\name{set1_unscaled_diff}
\alias{set1_unscaled_diff}
\docType{data}
\title{
The unstandardized F-statistics and ECF-statistics of simulated dataset 1
}
\description{
Generated by the function of "diff_gen" which performs comparative 
analysis of differential gene expression patterns between the simulated
case dataset1 and control dataset
}
\usage{data(set1_unscaled_diff)}
\format{
  The format is:

List of 2

 $ : num [1:500] 0.251 0.291 0.398 0.383 1.312 ...
 $ : num [1:500, 1:500] 0 0.03702 0.00924 0.01146 0.04543 ...

}
\details{
It is a list of length 2, which includes a vector of 
the F-statistics and a matrix of the ECF-statistics.
}
\examples{
data(set1_unscaled_diff)
}
\keyword{datasets}
