\name{simulated_data}
\alias{simulated_data}
\docType{data}
\title{
The simulated data sets used in the paper
}
\description{
It contains case data set1 to set6, as well as a common
control dataset
}
\usage{data(simulated_data)}
\format{
  The format is:
List of 7
 $ : num [1:20, 1:500] 0.00324 0.48968 0.51217 -0.70392 -0.72268 ...
  ..- attr(*, "dimnames")=List of 2

 $ : num [1:20, 1:500] 0.3621 -0.00548 1.83257 0.01457 -1.0223 ...
  ..- attr(*, "dimnames")=List of 2

 $ : num [1:20, 1:500] 0.3842 1.4062 2.5474 -0.0579 0.0179 ...
  ..- attr(*, "dimnames")=List of 2

 $ : num [1:20, 1:500] 0.278 -0.36 0.606 -1.37 0.853 ...
  ..- attr(*, "dimnames")=List of 2

 $ : num [1:20, 1:500] 2.53 1.092 1.937 0.822 -1.073 ...
  ..- attr(*, "dimnames")=List of 2

 $ : num [1:20, 1:500] 0.3621 -0.00548 1.83257 0.01457 -1.0223 ...
  ..- attr(*, "dimnames")=List of 2

 $ : num [1:20, 1:500] -0.754 0.586 0.487 -0.16 -0.975 ...
  ..- attr(*, "dimnames")=List of 2

}

\examples{
data(simulated_data)
}

\keyword{datasets}
