\name{COSINE-package}
\alias{COSINE-package}
\alias{COSINE}
\docType{package}
\title{
COndition SpecIfic subNEtwork identification
}
\description{This is a package to identify the single globally optimal
subnetwork which differs the most between two or more datasets}
\details{
\tabular{ll}{
Package: \tab COSINE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-07\cr
License: \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}
install.packages("COSINE")
}
\author{
Haisu Ma
Maintainer: Haisu Ma <haisu.ma@yale.edu>
}